Get rid of `parse_str_raw`
This trait method inherited from `serde-json` is currently unused, and
probably doesn't make sense in `lexpr`. The story behind the decision
to remove it is this:

In `serde-json`, `parse_str_raw` is used to parse strings without
UTF-8 validation, when a string is encountered in a position that
accepts a byte array (see the extensive documentation for
[`deserialize_bytes`]). This feature can be used to look at the raw
bytes in string literal, with the gotcha that should the string
contain JSON `\u` unicode escapes, the codepoints referred to by those
would end up UTF-8 encoded, thus resulting in a mix of encodings when
a non-UTF-8-encoded string contains `\u` escaped non-ASCII characters.

Considering that the whole parsing code in `lexpr` assumes an
ASCII-compatible encoding, there is syntax to encode all unicode
strings as ASCII S-expression text (using the R6RS/R7RS hex escapes),
and UTF-8 is (hopefully, by now), the prevalent ASCII-compatible
encoding, and the possible confusion when using another
ASCII-compatible encoding (such as latin-1) in combination with
escaped unicode codepoints (as mentioned above, this is also an issue
in `serde-json`), it does not seem worth keeping this feature.

[`deserialize_bytes`]: https://docs.serde.rs/serde_json/struct.Deserializer.html#method.deserialize_bytes