Fix parser for unidiomatic space-missing input
Several characters would not indicate the end of a symbol in the
parser; let's fix that by adding characters that may definitely not be
part of a symbol.

Note that this minimal fix means we still accept symbols that are
forbidden by Scheme or Emacs Lisp; fixing that is left for a future
commit.

Fixes: #40