Improve efficiency of `DeepSafeDrop::take_next_child_at_pos_index`
For `Value::Vector`. This operation is sometimes repeated up to
`vec.len()` times, which multiplies the computational complexity.  For
some `Value`-tree shapes with many longer vectors, this improvement
should help the performance of dropping via `DeepSafeValueDropper` a
lot so that the computational complexity is not worse than not having
the custom `Drop` impl.